chore: move big example to a test
